+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **Script/Config Changes**
 - Remove wood chip crushing wheel recipes
+- Return bucket to player when concrete bucket is drank & set stack size to 1
 
 **Mod Changes**
 - jamlib-forge-1.3.5+1.20.1 -> jamlib-forge-1.3.6+1.20.1-patch.1